+++ Source/BeatBloks.h
@@ -73,7 +73,7 @@ public:
 
    void CheckboxUpdated(Checkbox* checkbox, double time) override;
    //IFloatSliderListener
-   void FloatSliderUpdated(FloatSlider* slider, float oldVal, double time) override;
+   void FloatSliderUpdated(FloatSlider* slider, double oldVal, double time) override;
    //IFloatSliderListener
    void IntSliderUpdated(IntSlider* slider, int oldVal, double time) override;
    //IDropdownListener
@@ -137,7 +137,7 @@ private:
 
    Sample* mSample{ nullptr };
 
-   float mVolume{ .6 };
+   double mVolume{ .6 };
    FloatSlider* mVolumeSlider{ nullptr };
    float* mWriteBuffer;
    bool mPlay{ false };
@@ -145,15 +145,15 @@ private:
    bool mLoop{ true };
    Checkbox* mLoopCheckbox{ nullptr };
    int mMeasureEarly{ 0 };
-   float mClipStart{ 0 };
+   double mClipStart{ 0 };
    FloatSlider* mClipStartSlider{ nullptr };
-   float mClipEnd{ 1 };
+   double mClipEnd{ 1 };
    FloatSlider* mClipEndSlider{ nullptr };
-   float mZoomStart{ 0 };
+   double mZoomStart{ 0 };
    FloatSlider* mZoomStartSlider{ nullptr };
-   float mZoomEnd{ 1 };
+   double mZoomEnd{ 1 };
    FloatSlider* mZoomEndSlider{ nullptr };
-   float mOffset{ 0 };
+   double mOffset{ 0 };
    FloatSlider* mOffsetSlider{ nullptr };
    int mNumBars{ 1 };
    IntSlider* mNumBarsSlider{ nullptr };
@@ -188,9 +188,9 @@ private:
    Blok* mLastPlayedRemixBlok{ nullptr };
    float mLastLookupPlayhead{ 0 };
    ClickButton* mClearRemixButton{ nullptr };
-   float mRemixZoomStart{ 0 };
+   double mRemixZoomStart{ 0 };
    FloatSlider* mRemixZoomStartSlider{ nullptr };
-   float mRemixZoomEnd;
+   double mRemixZoomEnd;
    FloatSlider* mRemixZoomEndSlider{ nullptr };
    bool mBlockMultiPlaceEngaged{ false };
