+++ Source/Compressor.h
@@ -121,7 +121,7 @@ public:
    std::string GetType() override { return "compressor"; }
 
    void CheckboxUpdated(Checkbox* checkbox, double time) override;
-   void FloatSliderUpdated(FloatSlider* slider, float oldVal, double time) override;
+   void FloatSliderUpdated(FloatSlider* slider, double oldVal, double time) override;
 
    bool IsEnabled() const override { return mEnabled; }
 
@@ -134,14 +134,14 @@ private:
       height = mHeight;
    }
 
-   float mMix{ 1 };
-   float mDrive{ 1 };
-   float mThreshold{ -24 };
-   float mRatio{ 4 };
-   float mAttack{ .1f };
-   float mRelease{ 100 };
-   float mLookahead{ 3 };
-   float mOutputAdjust{ 1 };
+   double mMix{ 1 };
+   double mDrive{ 1 };
+   double mThreshold{ -24 };
+   double mRatio{ 4 };
+   double mAttack{ .1f };
+   double mRelease{ 100 };
+   double mLookahead{ 3 };
+   double mOutputAdjust{ 1 };
    FloatSlider* mMixSlider{ nullptr };
    FloatSlider* mDriveSlider{ nullptr };
    FloatSlider* mThresholdSlider{ nullptr };
